Fix various bugs
